fix the issue where user cannot send messages
